+++ packages/menu-service/src/types.ts
@@ -1,0 +1,8 @@
+export interface IMenuData {
+  name: string;
+  path: string;
+  icon?: string;
+  children?: IMenuData[];
+}
+
+export type MenuType = 'headerMenuConfig' | 'asideMenuConfig';