+++ extensions/iceworks-project-creator/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## 0.3.13
+
+- fix: if there is an active window, then do not show the welcome page
+
+## 0.3.12
+
+- fix: show welcome page many times
+- fix: refresh material did not clean cache
+
+## 0.3.11
+
+- feat: Add welcome page
+
 ## 0.3.10
 
 - feat: add Rax scaffolds to external