+++ extensions/iceworks-time-steward/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 0.1.4
+
+- fix: if there is an active window, then do not show the welcome page
+
+## 0.1.3
+
+- feat: Add welcome page
+
+## 0.1.2
+
+- refactor: just record PV
+
+## 0.1.1
+
+- fix: user info error
+
 ## 0.1.0
 
 - Initial release